requirements.txt
Flask==2.3.3
pycryptodome==3.18.0